switched to i32 for unification

+++ src/components/tiles.rs
@@ -1,7 +1,7 @@
 use amethyst::ecs::{NullStorage, prelude::{DenseVecStorage, Component}};
 use super::BuildingType;
 pub struct Tiles{ 
-    pub player: i8,
+    pub player: i32,
     pub buildingtype: Option<BuildingType>,
     pub x: i32,
     pub y: i32,
